pyproject.toml
[tool.poetry]
name = "filmes"
version = "0.1.0"
description = ""
authors = ["onesantos <wanderss@live.com>"]

[tool.poetry.dependencies]
python = "^3.8"
marshmallow-sqlalchemy = "^0.25.0"
Flask = "^2.0.0"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
flask-expects-json = "^1.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
